requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
pyaudio
py3-tts
#speech_recognition
python-dotenv
gtts